requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
dnspython==2.4.2
easyocr==1.7.1
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
fsspec==2023.12.2
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lazy_loader==0.3
Mako==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-bidi==0.4.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
sqlmodel==0.0.14
starlette==0.32.0.post1
sympy==1.12
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
